requirements.txt
arrow==0.17.0
asgiref==3.3.1
binaryornot==0.4.4
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
Django==3.1.4
django-model-utils==4.1.1
idna==2.10
Jinja2==2.11.2
jinja2-time==0.2.0
MarkupSafe==1.1.1
poyo==0.5.0
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2020.4
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.2
